// src/api/generated/models/apiclan-db.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { APIPlayerModel } from './apiplayer-model';

/**
 * 
 * @export
 * @interface APIClanDB
 */
export interface APIClanDB {
    /**
     * 
     * @type {string}
     * @memberof APIClanDB
     */
    'region': string;
    /**
     * 
     * @type {string}
     * @memberof APIClanDB
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof APIClanDB
     */
    'clan_id': number;
    /**
     * 
     * @type {string}
     * @memberof APIClanDB
     */
    'tag': string;
    /**
     * 
     * @type {number}
     * @memberof APIClanDB
     */
    'members_count': number;
    /**
     * 
     * @type {Array<APIPlayerModel>}
     * @memberof APIClanDB
     */
    'members': Array<APIPlayerModel>;
    /**
     * 
     * @type {number}
     * @memberof APIClanDB
     */
    'timestamp'?: number;
}