// src/api/generated/models/apirest-user-db.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { APIRegion } from './apiregion';

/**
 * 
 * @export
 * @interface APIRestUserDB
 */
export interface APIRestUserDB {
    /**
     * 
     * @type {APIRegion}
     * @memberof APIRestUserDB
     */
    'region': APIRegion;
    /**
     * 
     * @type {string}
     * @memberof APIRestUserDB
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof APIRestUserDB
     */
    'player_id': number;
}