// src/api/generated/models/apiregion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const APIRegion = {
    Eu: 'eu',
    Asia: 'asia',
    Com: 'com',
    Na: 'na'
} as const;

export type APIRegion = typeof APIRegion[keyof typeof APIRegion];